ENH: improved ufunc interface, now use x.ufunc.sin
