+++ odl/util/__init__.py
@@ -36,3 +36,5 @@ __all__ += phantom.__all__
 from . import graphics
 from .graphics import *
 __all__ += graphics.__all__
+
+from . import ufuncs
